+++ spec/unpoly/form_watch_fn_spec.js
@@ -15,6 +15,25 @@ extendDescribe('up.form', function() {
 
       describe('custom form fields', function() {
 
+        it('watches a form-associated custom element that has no { name } property', async function() {
+          const [, field] = htmlFixtureList(`
+            <form>
+              <test-attribute-named-field name="email" value="old@example.com"></test-attribute-named-field>
+            </form>
+          `)
+          const callback = jasmine.createSpy('watch callback')
+
+          up.watch(field, callback)
+
+          field.value = 'new@example.com'
+          Trigger.change(field)
+          await wait()
+
+          expect(callback.calls.count()).toBe(1)
+          expect(callback.calls.argsFor(0)[0]).toBe('new@example.com')
+          expect(callback.calls.argsFor(0)[1]).toBe('email')
+        })
+
         it('watches a form-associated custom element without configuration', async function() {
           const [, field] = htmlFixtureList(`
             <form>